% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV file into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file you want to process}
}
\value{
The function returns the content of the file in a tibble
}
\description{
This is a simple function which first reads 'filename' file in the working directory and
then outputs it as a tibble. You can input filename (using the \code{filename} argument).
It the file does not exist in the working directory, the function will stop and print an
error message; otherwise the function will call 'read_csv' function in "readr" package and
'tbl_df' function in "dplyr" package to read and transform the file into a tibble.
}
\examples{
\\donotrun{
fars_read("accident_2013.csv.bz2")
}

}
