% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read CSV file and Filter data by Month and Year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years in list you want to process}
}
\value{
a tibble with columns 'MONTH' and 'year'
}
\description{
The function will read files in the form "accident_'year'.csv.bz2" where 'year' is inputted
by user, read the data as a tibble and then select columns 'MONTH' and 'year' as output.
If file "accident_'year'.csv.bz2" does not exist, it will print a warning message and return
NULL. You can input the year using \code{list(years)} argument.
}
\examples{
years <- list(2013,2014)
fars_read_years(years)

}
