% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read CSV, Filter data by MONTH and year, and Summarize its counts}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years in list you want to process}
}
\value{
a tibble with columns 'MONTH' and 'years'
}
\description{
This function calls fars_read_years to read files in "accident_'year'.csv.bz2" form and
outputs a tibble with column 'MONTH' and 'year' first. Then it unlists (MONTH, year) pairs,
groups them, and counts their numbers. Finally, it "reshapes" the output as a tibble
with "MONTH", "years" form where "years" is the list in the input.
You can input the year using \code{list(years)} argument.
}
\examples{
years <- list(2013,2014)
fars_summarize_years(years)

}
