% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a character as "accident_'year'.csv.bz2"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year you want to process}
}
\value{
The function returns a character in the form "accident_'year'.csv.bz2"
}
\description{
The function will transfer the year you input into an integer and output as a character
named "accident_'year'.csv.bz2". You can input the year using \code{year} argument.
}
\examples{
\\donotrun{
make_filename(2013)
}


}
